# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthCheck AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CreateAssessmentRequestGender =
  (typeof CreateAssessmentRequestGender)[keyof typeof CreateAssessmentRequestGender];

export const CreateAssessmentRequestGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export type LifestyleDataSmokingStatus =
  (typeof LifestyleDataSmokingStatus)[keyof typeof LifestyleDataSmokingStatus];

export const LifestyleDataSmokingStatus = {
  never: "never",
  former: "former",
  current: "current",
} as const;

export type LifestyleDataAlcoholConsumption =
  (typeof LifestyleDataAlcoholConsumption)[keyof typeof LifestyleDataAlcoholConsumption];

export const LifestyleDataAlcoholConsumption = {
  none: "none",
  moderate: "moderate",
  heavy: "heavy",
} as const;

export type LifestyleDataExerciseFrequency =
  (typeof LifestyleDataExerciseFrequency)[keyof typeof LifestyleDataExerciseFrequency];

export const LifestyleDataExerciseFrequency = {
  sedentary: "sedentary",
  light: "light",
  moderate: "moderate",
  active: "active",
  very_active: "very_active",
} as const;

export type LifestyleDataDietType =
  (typeof LifestyleDataDietType)[keyof typeof LifestyleDataDietType];

export const LifestyleDataDietType = {
  omnivore: "omnivore",
  vegetarian: "vegetarian",
  vegan: "vegan",
  keto: "keto",
  mediterranean: "mediterranean",
  other: "other",
} as const;

export interface LifestyleData {
  smokingStatus: LifestyleDataSmokingStatus;
  alcoholConsumption: LifestyleDataAlcoholConsumption;
  exerciseFrequency: LifestyleDataExerciseFrequency;
  sleepHours?: number;
  /**
   * @minimum 1
   * @maximum 10
   */
  stressLevel?: number;
  dietType?: LifestyleDataDietType;
}

export interface CreateAssessmentRequest {
  /**
   * @minimum 1
   * @maximum 120
   */
  age: number;
  gender: CreateAssessmentRequestGender;
  /** Weight in kg */
  weight: number;
  /** Height in cm */
  height: number;
  symptoms?: string[];
  lifestyle?: LifestyleData;
  medicalHistory?: string[];
}

export type AssessmentResultRiskLevel =
  (typeof AssessmentResultRiskLevel)[keyof typeof AssessmentResultRiskLevel];

export const AssessmentResultRiskLevel = {
  low: "low",
  moderate: "moderate",
  high: "high",
  critical: "critical",
} as const;

export interface CategoryScores {
  physical: number;
  mental: number;
  nutrition: number;
  sleep: number;
  lifestyle: number;
}

export interface AssessmentResult {
  id: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  overallScore: number;
  bmi: number;
  bmiCategory: string;
  riskLevel: AssessmentResultRiskLevel;
  recommendations: string[];
  categoryScores: CategoryScores;
  createdAt: string;
  age: number;
  gender: string;
  weight: number;
  height: number;
}

export interface CreateVitalsRequest {
  /** Heart rate in bpm */
  heartRate: number;
  /** Systolic blood pressure */
  systolicBP?: number;
  /** Diastolic blood pressure */
  diastolicBP?: number;
  /** Temperature in Celsius */
  temperature?: number;
  /** SpO2 percentage */
  oxygenSaturation?: number;
  /** Blood glucose in mg/dL */
  bloodGlucose?: number;
}

export interface VitalsEntry {
  id: number;
  heartRate: number;
  systolicBP?: number;
  diastolicBP?: number;
  temperature?: number;
  oxygenSaturation?: number;
  bloodGlucose?: number;
  recordedAt: string;
}

export type HealthTipCategory =
  (typeof HealthTipCategory)[keyof typeof HealthTipCategory];

export const HealthTipCategory = {
  nutrition: "nutrition",
  exercise: "exercise",
  sleep: "sleep",
  mental_health: "mental_health",
  hydration: "hydration",
  prevention: "prevention",
} as const;

export interface HealthTip {
  id: number;
  category: HealthTipCategory;
  title: string;
  content: string;
  icon: string;
}