# lib/api-zod/src/generated/types/assessmentResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthCheck AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentResultRiskLevel } from "./assessmentResultRiskLevel";
import type { CategoryScores } from "./categoryScores";

export interface AssessmentResult {
  id: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  overallScore: number;
  bmi: number;
  bmiCategory: string;
  riskLevel: AssessmentResultRiskLevel;
  recommendations: string[];
  categoryScores: CategoryScores;
  createdAt: Date;
  age: number;
  gender: string;
  weight: number;
  height: number;
}