# lib/api-zod/src/generated/types/categoryScores.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthCheck AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryScores {
  physical: number;
  mental: number;
  nutrition: number;
  sleep: number;
  lifestyle: number;
}