# lib/api-zod/src/generated/types/createAssessmentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthCheck AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAssessmentRequestGender } from "./createAssessmentRequestGender";
import type { LifestyleData } from "./lifestyleData";

export interface CreateAssessmentRequest {
  /**
   * @minimum 1
   * @maximum 120
   */
  age: number;
  gender: CreateAssessmentRequestGender;
  /** Weight in kg */
  weight: number;
  /** Height in cm */
  height: number;
  symptoms?: string[];
  lifestyle?: LifestyleData;
  medicalHistory?: string[];
}