# lib/api-zod/src/generated/types/lifestyleData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthCheck AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LifestyleDataAlcoholConsumption } from "./lifestyleDataAlcoholConsumption";
import type { LifestyleDataDietType } from "./lifestyleDataDietType";
import type { LifestyleDataExerciseFrequency } from "./lifestyleDataExerciseFrequency";
import type { LifestyleDataSmokingStatus } from "./lifestyleDataSmokingStatus";

export interface LifestyleData {
  smokingStatus: LifestyleDataSmokingStatus;
  alcoholConsumption: LifestyleDataAlcoholConsumption;
  exerciseFrequency: LifestyleDataExerciseFrequency;
  sleepHours?: number;
  /**
   * @minimum 1
   * @maximum 10
   */
  stressLevel?: number;
  dietType?: LifestyleDataDietType;
}